+++ tests/elf_binary_printer_test.py
@@ -9,7 +9,11 @@ import gtirb_test_helpers as gth
 import dummyso
 import hello_world
 
-from pprinter_helpers import BinaryPPrinterTest, run_asm_pprinter
+from pprinter_helpers import (
+    BinaryPPrinterTest,
+    run_asm_pprinter,
+    run_asm_pprinter_with_version_script,
+)
 
 
 @unittest.skipUnless(os.name == "posix", "only runs on Linux")
@@ -227,24 +231,27 @@ class ElfBinaryPrinterTests(BinaryPPrinterTest):
             # Just verify binary_print succeeded.
             pass
 
-    def test_base_version(self):
+    def test_dummyso_version_script(self):
         """
-        Make sure that base version is not printed out
+        Test printing version script
         """
-        ir, module, bi = self.build_basic_ir()
+        ir, module = self.build_basic_ir()
 
-        foo_block = gth.add_code_block(bi, b"\xC3")
-        foo_uuid = gth.add_function(module, "foo", foo_block)
-
-        # Get the foo function symbol
-        symbol_foo = module.aux_data["functionNames"].data[foo_uuid]
-
-        # Create another symbol pointing to the same block
-        symbol_foo2 = gth.add_symbol(module, "foo", foo_block)
-        module.aux_data["elfSymbolInfo"].data[symbol_foo2.uuid] = (
+        proxy_foo = gth.add_proxy_block(module)
+        symbol_foo = gth.add_symbol(module, "foo", proxy_foo)
+        module.aux_data["elfSymbolInfo"].data[symbol_foo.uuid] = (
             0,
             "FUNC",
             "GLOBAL",
+            "DEFAULT",
+            0,
+        )
+        proxy_bar = gth.add_proxy_block(module)
+        symbol_bar = gth.add_symbol(module, "bar", proxy_bar)
+        module.aux_data["elfSymbolInfo"].data[symbol_bar.uuid] = (
+            0,
+            "FUNC",
+            "LOCAL",
             "DEFAULT",
             0,
         )
@@ -256,31 +263,79 @@ class ElfBinaryPrinterTests(BinaryPPrinterTest):
             ),
             data=(
                 # ElfSymVerDefs
-                {
-                    1: (["LIBA_1.0"], 0),
-                    # Flags=1: base version
-                    2: (["libmya.so"], 1),
-                },
+                {1: (["LIBA_1.0"], 0), 2: (["LIBA_2.0"], 0)},
                 # ElfSymVerNeeded
-                {"libmya.so": {1: "LIBA_1.0"}},
+                {"libmya.so": {1: "LIBA_1.0", 2: "LIBA_2.0"}},
                 # ElfSymbolVersionsEntries
                 {
                     symbol_foo.uuid: (1, False),
-                    # symbol_foo2 gets the base version
-                    symbol_foo2.uuid: (2, False),
+                    symbol_bar.uuid: (2, False),
                 },
             ),
         )
 
-        asm = run_asm_pprinter(ir)
+        vs = run_asm_pprinter_with_version_script(ir)
 
-        self.assertRegexMatch(asm, r"foo@@LIBA_1.0")
-        # The base version should not pr
-        self.assertNotRegex(
-            asm,
-            r"foo@@libmya.so",
-            msg="The base version 'libmya.so' should not be printed out",
-        )
+        pattern1 = r"LIBA_1.0\s+{\s+global:\s+foo;\s+local:\s+\*;\s+};"
+        self.assertRegexMatch(vs, pattern1)
+        pattern2 = r"LIBA_2.0\s+{\s+local:\s+\*;\s+};"
+        self.assertRegexMatch(vs, pattern2)
+        self.assertTrue("bar" not in vs)
+
+        def test_base_version(self):
+            """
+            Make sure that base version is not printed out
+            """
+            ir, module, bi = self.build_basic_ir()
+
+            foo_block = gth.add_code_block(bi, b"\xC3")
+            foo_uuid = gth.add_function(module, "foo", foo_block)
+
+            # Get the foo function symbol
+            symbol_foo = module.aux_data["functionNames"].data[foo_uuid]
+
+            # Create another symbol pointing to the same block
+            symbol_foo2 = gth.add_symbol(module, "foo", foo_block)
+            module.aux_data["elfSymbolInfo"].data[symbol_foo2.uuid] = (
+                0,
+                "FUNC",
+                "GLOBAL",
+                "DEFAULT",
+                0,
+            )
+            module.aux_data["elfSymbolVersions"] = gtirb.AuxData(
+                type_name=(
+                    "tuple<mapping<uint16_t,tuple<sequence<string>,uint16_t>>,"
+                    "mapping<string,mapping<uint16_t,string>>,"
+                    "mapping<UUID,tuple<uint16_t,bool>>>"
+                ),
+                data=(
+                    # ElfSymVerDefs
+                    {
+                        1: (["LIBA_1.0"], 0),
+                        # Flags=1: base version
+                        2: (["libmya.so"], 1),
+                    },
+                    # ElfSymVerNeeded
+                    {"libmya.so": {1: "LIBA_1.0"}},
+                    # ElfSymbolVersionsEntries
+                    {
+                        symbol_foo.uuid: (1, False),
+                        # symbol_foo2 gets the base version
+                        symbol_foo2.uuid: (2, False),
+                    },
+                ),
+            )
+
+            asm = run_asm_pprinter(ir)
+
+            self.assertRegexMatch(asm, r"foo@@LIBA_1.0")
+            # The base version should not pr
+            self.assertNotRegex(
+                asm,
+                r"foo@@libmya.so",
+                msg="The base version 'libmya.so' should not be printed out",
+            )
 
     def test_use_gcc(self):
         """